Fix compiling int and guard patterns
Integer patterns could break certain nested patterns, such as
`(4, (5, 7)) -> A` due to how fallback rows were handled: they were only
processed in the fallback case of the outer pattern, resulting in the
inner pattern not having enough/the expected rows to process. This in
turn could result in exhaustive matches being treated as non-exhaustive.
For example, prior to this commit a match like this was considered
non-exhaustive:

    match (10, 20) {
      case (10, 20) -> A
      case (_, _)   -> B
    }

As part of this commit we also fix guard patterns, which were broken
when the pattern was the same. For example, matches like this weren't
handled correctly:

    match 42 {
      case 42 if x -> A
      case 42 if y -> B
      case 42      -> C
      ...
    }

This was caused by the compiler keeping track of the integer columns
already processed, without taking into account guard expressions. This
would result in `case 42 if y` and `case 42` being skipped.

To fix this, when compiling integer patterns we track the index of each
integer value processed. When we encounter a duplicate integer, we just
push the current row into that integer's existing list of rows.